+++ src/utilities/getDidCreationDetails/getDidCreationDetails.ts
@@ -2,28 +2,17 @@ import { BlockchainApiConnection } from '@kiltprotocol/chain-helpers';
 
 import { FullDidCreationBuilder } from '@kiltprotocol/did';
 
-import ky from 'ky';
+import { HexString } from '@polkadot/util/types';
 
 import {
   getKeystoreFromSeed,
   getLightDidFromSeed,
 } from '../identities/identities';
 import { getDidEncryptionKey, parseDidUri } from '../did/did';
-import { getBackendEndpoints } from '../getBackendEndpoints/getBackendEndpoints';
 
-export async function getPromoStatus(): Promise<{
-  account: string;
-  remaining_dids: number;
-  is_active: boolean;
-}> {
-  const backendEndpoints = await getBackendEndpoints();
-
-  return await ky.get(backendEndpoints.promoStatus).json();
-}
-
-interface CreationDetails {
-  creationDetails: string;
-  signature: string;
+export interface CreationDetails {
+  creationDetails: HexString;
+  signature: HexString;
   did: string;
 }
 
@@ -51,23 +40,4 @@ export async function getDidCreationDetails(
     signature: extrinsic.args[1].toHex(),
     did,
   };
-}
-
-export async function createDid(
-  input: Omit<CreationDetails, 'did'>,
-): Promise<{ tx_hash: string }> {
-  const backendEndpoints = await getBackendEndpoints();
-
-  return ky.post(backendEndpoints.createDid, { json: input }).json();
-}
-
-export async function waitFinalized(tx_hash: string): Promise<boolean> {
-  const backendEndpoints = await getBackendEndpoints();
-
-  return ky
-    .get(backendEndpoints.waitFinalized, {
-      searchParams: { tx_hash },
-      timeout: false,
-    })
-    .json();
 }
