+++ src/views/DidUpgradePromo/DidUpgradePromo.test.tsx
@@ -3,11 +3,11 @@ import { act, render } from '../../testing/testing';
 import { identitiesMock as identities } from '../../utilities/identities/IdentitiesProvider.mock';
 import { waitForGetPassword } from '../../channels/SavedPasswordsChannels/SavedPasswordsChannels.mock';
 
-import { getPromoStatus } from '../../utilities/didUpgradePromo/didUpgradePromo';
+import { getPromoStatus } from '../../utilities/promoBackend/promoBackend';
 
 import { DidUpgradePromo } from './DidUpgradePromo';
 
-jest.mock('../../utilities/didUpgradePromo/didUpgradePromo');
+jest.mock('../../utilities/promoBackend/promoBackend');
 
 jest.mocked(getPromoStatus).mockResolvedValue({
   account: '4oY2qsDpYBf2LqahCTmEC4iudf667CRT3iNoBmMLfznZoGcM',
